requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
asyncpg==0.30.0
attrs==25.4.0
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
dataclasses-json==0.6.7
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
deepl==1.24.0
defusedxml==0.7.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
docutils==0.22.3
executing==2.2.1
fastapi==0.121.0
fastjsonschema==2.21.2
feedparser==6.0.12
filelock==3.20.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
id==1.5.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
instructor==1.13.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jpype1==1.6.0
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keyring==25.6.0
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
konlpy==0.6.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-mcp-adapters==0.1.12
langchain-openai==0.3.35
langchain-postgres==0.0.16
langchain-teddynote==0.5.4
langchain-text-splitters==0.3.11
langchain-upstage==0.7.4
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.41
lark==1.3.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mcp==1.21.0
mdurl==0.1.2
mistune==3.1.4
mmh3==4.1.0
more-itertools==10.8.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.2
nltk==3.9.2
nodeenv==1.9.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.4
olefile==0.47
openai==2.7.1
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pdf2image==1.17.0
pgvector==0.3.6
pillow==12.0.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.11.0
platformdirs==4.5.0
pre_commit==4.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg-pool==3.2.7
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ragas==0.3.8
rank-bm25==0.2.2
readme_renderer==44.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
scikit-network==0.33.3
scipy==1.16.3
Send2Trash==1.8.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
tavily-python==0.7.12
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.20.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
twine==6.2.0
ty==0.0.1a25
typer==0.20.0
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
